chore(mobile): Makes Icon component comply with pods convention
